Remove server echo to client updating room topic
Fixing an issue where ping between client and server caused updates
being typed by the user to be overwritten by the server sending older
updates back to the client and so on.

Now just notify only other users in the room that the title has been
updated, and use the client copy as the canonical version on the latest
updating client.